# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Football API - AYO Technical Test 2026",
    "description": "Backend API untuk manajemen tim sepakbola amatir perusahaan XYZ. Menyediakan CRUD untuk Tim, Pemain, Pertandingan, pencatatan Gol, dan Laporan hasil pertandingan.\n\nSemua endpoint GET bersifat publik. Endpoint POST/PUT/DELETE memerlukan header `X-API-Key` (lihat Authorize di kanan atas).",
    "version": "1.0",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header",
      "description": "API Key untuk endpoint yang membutuhkan otorisasi (create/update/delete)."
    }
  },
...
  "paths": {
    "/teams": {
      "get": {
        "summary": "Daftar semua tim",
        "tags": [
          "Teams"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeamListResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Tambah tim baru",
        "tags": [
          "Teams"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTeamRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/TeamResponse"
            }
          },
          "400": {
            "description": "Bad Request - validasi gagal",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized - API key tidak valid",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict - nama tim sudah ada",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/teams/{id}": {
      "get": {
        "summary": "Detail tim",
        "tags": [
          "Teams"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID tim"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeamResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Update tim",
        "tags": [
          "Teams"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID tim"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/UpdateTeamRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeamResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict - nama tim sudah ada",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Hapus tim (soft delete)",
        "tags": [
          "Teams"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID tim"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeletedResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Team": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Persib Bandung"
        },
        "logo": {
          "type": "string",
          "example": "https://example.com/logo.png"
        },
        "founded": {
          "type": "integer",
          "example": 1933
        },
        "headquarters_address": {
          "type": "string",
          "example": "Jl. Sulanjana No.17"
        },
        "headquarters_city": {
          "type": "string",
          "example": "Bandung"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "CreateTeamRequest": {
      "type": "object",
      "required": [
        "name",
        "founded",
        "headquarters_address",
        "headquarters_city"
      ],
      "properties": {
        "name": {
          "type": "string",
          "minLength": 2,
          "maxLength": 100,
          "example": "Persib Bandung"
        },
        "logo": {
          "type": "string",
          "example": "https://example.com/logo.png"
        },
        "founded": {
          "type": "integer",
          "minimum": 1800,
          "maximum": 2100,
          "example": 1933
        },
        "headquarters_address": {
          "type": "string",
          "example": "Jl. Sulanjana No.17"
        },
        "headquarters_city": {
          "type": "string",
          "example": "Bandung"
        }
      }
    },
    "UpdateTeamRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "minLength": 2,
          "maxLength": 100
        },
        "logo": {
          "type": "string"
        },
        "founded": {
          "type": "integer",
          "minimum": 1800,
          "maximum": 2100
        },
        "headquarters_address": {
          "type": "string"
        },
        "headquarters_city": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "request failed"
        },
        "error": {
          "type": "string",
          "example": "terjadi kesalahan"
        }
      }
    },
    "TeamResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        },
        "data": {
          "$ref": "#/definitions/Team"
        }
      }
    },
    "TeamListResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Team"
          }
        }
      }
    },
...
    "DeletedResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "OK"
        }
      }
    }
  }
}